requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
google-generativeai==0.3.2
telethon==1.33.1
Pillow==10.1.0
python-dateutil==2.8.2
SQLAlchemy==2.0.25
pydantic==2.5.3
loguru==0.7.2
httpx==0.26.0 